+++ CHANGELOG.md
@@ -16,3 +16,13 @@
 ## 0.0.4
 
 - fix: Chinese be transformed to empty.
+
+## 0.0.5
+
+- feat: can select multiple cases.
+- feat: add Capital Case.
+- feat: add path/case.
+
+## 0.0.6
+
+- feat: start with last searched value or current editor selection